Switch to qtpy and PyQt.
PyQt6 is the only Qt binding I could get to install on ARM
Mac. Unfortunately it's not supported by Qt.py (which has no releases
since 2020 so it may be abandoned). So I had to switch to qtpy.

But PyQt6 won't install on my Linux machine. So use PyQt5 in the
Pipfile -- on Mac the virtualenv will need to be fixed up manually.

The run script sets an environment variable telling qtpy to use PyQt6
on Mac and PyQt5 elsewhere.